Revert "Mongodb in memory testing"
