+++ .openshift/config/server.xml
@@ -41,7 +41,7 @@
     -->
     <Resource name="UserDatabase" auth="Container"
               type="org.apache.catalina.UserDatabase"
-              description="org.utn.marvellator.model.User database that can be updated and saved"
+              description="marvellator.User database that can be updated and saved"
               factory="org.apache.catalina.users.MemoryUserDatabaseFactory"
               pathname="conf/tomcat-users.xml" />
   </GlobalNamingResources>
