+++ README.md
@@ -1,19 +1,3 @@
-# Marvellator
-
-Prerequisites
-* Mongo installed ([steps](http://docs.mongodb.org/manual/installation/))
-* Run mongod in a separate shell
-
-Run project:
-
-```
-$ mvn clean install
-$ mvn spring-boot:run
-```
-Then navigate to localhost:8080
-
-------
-
 The OpenShift `jbossews` cartridge documentation can be found at:
 
 http://openshift.github.io/documentation/oo_cartridge_guide.html#tomcat
