+++ src/main/java/marvellator/Application.java
@@ -1,4 +1,4 @@
-package org.utn.marvellator;
+package marvellator;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,7 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class Application {
 
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-}
+        public static void main(String[] args) {
+            SpringApplication.run(Application.class, args);
+        }
+    }
