+++ src/main/java/marvellator/CharacterController.java
@@ -1,14 +1,29 @@
-package org.utn.marvellator.controller;
+package marvellator;
 
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
-@Controller
+@RestController
 public class CharacterController {
+    /* API */
+    @RequestMapping(value = "api/characters", method = RequestMethod.GET)
+    public String characters() {
+        return "List of characters \n";
+    }
 
+    @RequestMapping(value = "api/character/{id}", method = RequestMethod.POST)
+    public String setAsFavorite(@PathVariable int id) {
+        return "Character set as favorite. Its id is: " + id + "\n";
+    }
+
+    @RequestMapping(value = "/api/ranking", method = RequestMethod.GET)
+    public String ranking() {
+        return "Ranking of users \n";
+    }
+
+    /* FRONT END */
     @RequestMapping(value = "/characters", method = RequestMethod.GET)
     public String charactersPage() {
         return "Characters Page \n";
