+++ src/main/java/marvellator/GroupController.java
@@ -1,4 +1,4 @@
-package org.utn.marvellator.controller.rest;
+package marvellator;
 
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -6,8 +6,17 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class GroupRestController {
+public class GroupController {
+    /* API */
+    @RequestMapping(value = "api/groups", method = RequestMethod.GET)
+    public String groups() {
+        return "List of groups \n";
+    }
 
+    @RequestMapping(value = "api/groups", method = RequestMethod.POST)
+    public String newGroup() {
+        return "New group \n";
+    }
 
     @RequestMapping(value = "api/groups/{id}", method = RequestMethod.GET)
     public String group(@PathVariable int id) {
@@ -24,15 +33,14 @@ public class GroupRestController {
         return "Group deleted \n";
     }
 
-    @RequestMapping(value = "api/groups", method = RequestMethod.GET)
-    public String groups() {
-        return "List of groups \n";
+    /* FRONT END */
+    @RequestMapping(value = "/groups", method = RequestMethod.GET)
+    public String groupsPage() {
+        return "Groups Page \n";
     }
 
-    @RequestMapping(value = "api/groups", method = RequestMethod.POST)
-    public String newGroup() {
-        return "New group \n";
+    @RequestMapping(value = "/groups/{id}", method = RequestMethod.GET)
+    public String groupPage(@PathVariable int id) {
+        return "A group Page \n";
     }
-
-
 }
