+++ src/main/java/marvellator/HomeController.java
@@ -1,14 +1,13 @@
-package org.utn.marvellator.controller;
+package marvellator;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
+@RestController
 public class HomeController {
-
     @RequestMapping(value = "/")
     public String home() {
-        return "home";
+        return "Home page \n";
     }
 }
