--- src/main/java/org/utn/marvellator/model/User.java
@@ -1,72 +1,0 @@
-package org.utn.marvellator.model;
-
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.index.Indexed;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Document(collection = "user")
-public class User {
-
-	@Id
-	private String id;
-
-	private String name;
-
-	@Indexed(unique = true)
-	private String userName;
-
-	private String password;
-
-	private String email;
-
-	public User(){
-	}
-
-	public User(String name, String userName, String password) {
-		this.name = name;
-		this.userName = userName;
-		this.password = password;
-	}
-
-
-	public User(String userName) {
-		this.userName = userName;
-	}
-
-	public String getId() {
-		return id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getUserName() {
-		return userName;
-	}
-
-	public void setUserName(String userName) {
-		this.userName = userName;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-}
