--- src/main/java/org/utn/marvellator/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package org.utn.marvellator.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.utn.marvellator.model.User;
-
-public interface UserRepository extends MongoRepository<User, String>{
-
-    User findFirstByUserName(String userName);
-
-}
