--- src/main/java/org/utn/marvellator/service/UserService.java
@@ -1,8 +1,0 @@
-package org.utn.marvellator.service;
-
-import org.utn.marvellator.model.User;
-
-public interface UserService {
-
-    void registerUser(User user);
-}
