--- src/main/resources/templates/characters.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Characters</title>
-</head>
-<body>
-
-</body>
-</html>
