--- src/main/resources/templates/favorites.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Favorites</title>
-</head>
-<body>
-
-</body>
-</html>
