--- src/main/resources/templates/groups.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Groups of characters</title>
-</head>
-<body>
-
-</body>
-</html>
