+++ src/main/resources/templates/register.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Ranking of characters ordered by times favorited</title>
+    <title>Register</title>
 </head>
 <body>
