--- src/main/resources/templates/signup.html
@@ -1,16 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>USERS</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-</head>
-<body>
-<h1>Create a new account <strike>and worship satan</strike></h1>
-<form action="#" th:action="@{/user}" method="post">
-    <p><label>Username</label><input type="text" value="userName"/></p>
-    <p><label>Name</label><input type="text" value="name"/></p>
-    <p><label>Password</label><input type="text" value="password"/></p>
-    <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
-</form>
-</body>
-</html>
