+++ src/main/resources/templates/user.html
@@ -5,8 +5,10 @@
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
 </head>
 <body>
-<h1>User registered successfully :)</h1>
-    <p>Hi, {user.name}, your user is {user.userName}</p>
-
+<h1>User!</h1>
+<form action="#" th:action="@{/user}" th:object="${user}" method="post">
+    <p th:text=" ${user.nickName} + ${user.userName}" />
+    <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
+</form>
 </body>
 </html>
