+++ src/main/webapp/WEB-INF/web.xml
@@ -24,3 +24,4 @@
         <url-pattern>*.form</url-pattern>
     </servlet-mapping>
 </web-app>
+
